Fixed bugs and separete components

+++ headers/abstractobjectdetectionmodel.h
@@ -14,7 +14,7 @@ public:
    * @return Status, class id, score. Class id and score can be equal -1, if
    * objects not detected or was problems in process.
    */
-  virtual std::map<int, float> forward(const QImage &image) = 0;
+  virtual std::map<int, double> forward(const QImage &image) = 0;
 
   /**
    * Enables GPU for the TensorflowModel.
